Pause YouTube videos when closing popup.
YouTube videos would continue to play in the background after
the containing feedback popup was closed.

This commit pauses all YouTube videos present in the popup
when the popup is closed.

To test:

1. Add image explorer block to a lesson in the CMS.
   The default content already includes a youtube video.
2. Click the hotspot with the youtube video and start the video.
3. Close the popup.

Without this patch, the video would continue playing in the background.
With this patch applied, the video will be paused when closing the popup.