build: Don't install an editable xblock-sdk
The only reason that we needed the xblock-sdk as editable is beacause
this repo was not explictily stating its test requirements and instead,
relying on getting the xblock-sdk dev requirements via `make install`.

I updated test_requirements.txt with a snapshot of the requirements
needed to get the tests passing withot the editable install.

This repo needs a lot more work but now it is doing one less sketchy
thing.